refactor: 시큐리티 구조 개선
